+++ workers/unity/Packages/com.improbable.gdk.core/Worker/DefaultWorkerConnector.cs
@@ -1,4 +1,6 @@
 using System;
+using Improbable.Worker.CInterop;
+using Alpha = Improbable.Worker.CInterop.Alpha;
 using UnityEngine;
 
 namespace Improbable.Gdk.Core
@@ -10,62 +12,169 @@ namespace Improbable.Gdk.Core
         /// </summary>
         public bool UseExternalIp;
 
-        /// <summary>
-        ///     Determines whether to connect via the locator.
-        /// </summary>
-        /// <returns>True, if should connect via the Locator, false otherwise.</returns>
-        protected override bool ShouldUseLocator()
+        protected override ConnectionService GetConnectionService()
         {
+            var commandLineArguments = Environment.GetCommandLineArgs();
+            var commandLineArgs = CommandLineUtility.ParseCommandLineArgs(commandLineArguments);
+
+            if (commandLineArgs.ContainsKey(RuntimeConfigNames.SteamDeploymentTag) ||
+                commandLineArgs.ContainsKey(RuntimeConfigNames.SteamTicket))
+            {
+                return ConnectionService.Locator;
+            }
+
+            if (commandLineArgs.ContainsKey(RuntimeConfigNames.LoginToken))
+            {
+                return commandLineArgs.ContainsKey(RuntimeConfigNames.PlayerIdentityToken)
+                    ? ConnectionService.AlphaLocator
+                    : ConnectionService.Locator;
+            }
+
+            return ConnectionService.Receptionist;
+        }
+
+        protected override ConnectionParameters GetConnectionParameters(string workerType, ConnectionService service)
+        {
+            // UseExternalIp needs to be true when using the locator
+            var useExternalIp = service == ConnectionService.Locator
+                || service == ConnectionService.AlphaLocator
+                || UseExternalIp;
+
             if (Application.isEditor)
             {
-                return false;
+                return new ConnectionParameters
+                {
+                    WorkerType = workerType,
+                    Network =
+                    {
+                        ConnectionType = RuntimeConfigDefaults.LinkProtocol,
+                        UseExternalIp = useExternalIp,
+                    },
+                    EnableProtocolLoggingAtStartup = false,
+                    DefaultComponentVtable = new ComponentVtable(),
+                };
             }
 
             var commandLineArguments = Environment.GetCommandLineArgs();
             var commandLineArgs = CommandLineUtility.ParseCommandLineArgs(commandLineArguments);
-            var shouldUseLocator = commandLineArgs.ContainsKey(RuntimeConfigNames.LoginToken) ||
-                commandLineArgs.ContainsKey(RuntimeConfigNames.SteamDeploymentTag) ||
-                commandLineArgs.ContainsKey(RuntimeConfigNames.SteamTicket);
-            return shouldUseLocator;
+            var linkProtocol = CommandLineUtility.GetCommandLineValue(
+                commandLineArgs, RuntimeConfigNames.LinkProtocol, RuntimeConfigDefaults.LinkProtocol);
+
+            return new ConnectionParameters
+            {
+                WorkerType = CommandLineUtility.GetCommandLineValue(
+                    commandLineArgs, RuntimeConfigNames.WorkerType, workerType),
+                Network =
+                {
+                    ConnectionType = linkProtocol,
+                    UseExternalIp = useExternalIp,
+                },
+                EnableProtocolLoggingAtStartup = false,
+                DefaultComponentVtable = new ComponentVtable()
+            };
         }
 
         protected override ReceptionistConfig GetReceptionistConfig(string workerType)
         {
-            ReceptionistConfig config;
-
             if (Application.isEditor)
             {
-                config = new ReceptionistConfig
+                return new ReceptionistConfig
                 {
-                    WorkerType = workerType,
-                    WorkerId = CreateNewWorkerId(workerType),
-                    UseExternalIp = UseExternalIp
+                    ReceptionistHost = RuntimeConfigDefaults.ReceptionistHost,
+                    ReceptionistPort = RuntimeConfigDefaults.ReceptionistPort,
+                    WorkerId = CreateNewWorkerId(workerType)
                 };
             }
-            else
-            {
-                var commandLineArguments = Environment.GetCommandLineArgs();
-                var commandLineArgs = CommandLineUtility.ParseCommandLineArgs(commandLineArguments);
-                config = ReceptionistConfig.CreateConnectionConfigFromCommandLine(commandLineArgs);
-                config.WorkerType = workerType;
-                config.UseExternalIp = UseExternalIp;
-                if (!commandLineArgs.ContainsKey(RuntimeConfigNames.WorkerId))
-                {
-                    config.WorkerId = CreateNewWorkerId(workerType);
-                }
-            }
 
-            return config;
+            var commandLineArguments = Environment.GetCommandLineArgs();
+            var commandLineArgs = CommandLineUtility.ParseCommandLineArgs(commandLineArguments);
+            return new ReceptionistConfig
+                {
+                    ReceptionistHost = CommandLineUtility.GetCommandLineValue(
+                        commandLineArgs, RuntimeConfigNames.ReceptionistHost, RuntimeConfigDefaults.ReceptionistHost),
+                    ReceptionistPort = CommandLineUtility.GetCommandLineValue(
+                        commandLineArgs, RuntimeConfigNames.ReceptionistPort, RuntimeConfigDefaults.ReceptionistPort),
+                    WorkerId = CommandLineUtility.GetCommandLineValue(
+                        commandLineArgs, RuntimeConfigNames.WorkerId, CreateNewWorkerId(workerType)),
+                };
         }
 
-        protected override LocatorConfig GetLocatorConfig(string workerType)
+        protected override LocatorConfig GetLocatorConfig()
         {
             var commandLineArguments = Environment.GetCommandLineArgs();
             var commandLineArgs = CommandLineUtility.ParseCommandLineArgs(commandLineArguments);
-            var config = LocatorConfig.CreateConnectionConfigFromCommandLine(commandLineArgs);
-            config.WorkerType = workerType;
-            config.WorkerId = CreateNewWorkerId(workerType);
-            return config;
+
+            var projectName = CommandLineUtility.GetCommandLineValue(
+                commandLineArgs, RuntimeConfigNames.ProjectName, string.Empty);
+
+            if (string.IsNullOrEmpty(projectName))
+            {
+                throw new ConnectionFailedException("Project name is not set. Can't connect via the Locator.", ConnectionErrorReason.InvalidConfig);
+            }
+
+            var loginToken = CommandLineUtility.GetCommandLineValue(
+                commandLineArgs, RuntimeConfigNames.LoginToken, string.Empty);
+            var steamDeploymentTag = CommandLineUtility.GetCommandLineValue(
+                commandLineArgs, RuntimeConfigNames.SteamDeploymentTag, string.Empty);
+            var steamTicket = CommandLineUtility.GetCommandLineValue(
+                commandLineArgs, RuntimeConfigNames.SteamTicket, string.Empty);
+
+
+            LocatorCredentialsType credentialType;
+            if (!string.IsNullOrEmpty(loginToken))
+            {
+                credentialType = LocatorCredentialsType.LoginToken;
+            }
+            else if (!string.IsNullOrEmpty(steamDeploymentTag) && !string.IsNullOrEmpty(steamTicket))
+            {
+                credentialType = LocatorCredentialsType.Steam;
+            }
+            else
+            {
+                throw new ConnectionFailedException("Neither steam credentials nor login token is set. Can't connect via the Locator.", ConnectionErrorReason.InvalidConfig);
+            }
+
+            return new LocatorConfig
+            {
+                DeploymentListCallback = SelectDeploymentName,
+                LocatorHost = CommandLineUtility.GetCommandLineValue(
+                    commandLineArgs, RuntimeConfigNames.LocatorHost, RuntimeConfigDefaults.LocatorHost),
+                LocatorParameters = new LocatorParameters
+                {
+                    CredentialsType = credentialType,
+                    LoginToken = new LoginTokenCredentials
+                    {
+                        Token = loginToken,
+                    },
+                    Steam = new SteamCredentials
+                    {
+                        DeploymentTag = steamDeploymentTag,
+                        Ticket = steamTicket,
+                    },
+                    ProjectName = projectName,
+                }
+            };
+        }
+
+        protected override AlphaLocatorConfig GetAlphaLocatorConfig()
+        {
+            var commandLineArguments = Environment.GetCommandLineArgs();
+            var commandLineArgs = CommandLineUtility.ParseCommandLineArgs(commandLineArguments);
+            return new AlphaLocatorConfig
+            {
+                LocatorHost = CommandLineUtility.GetCommandLineValue(
+                    commandLineArgs, RuntimeConfigNames.LocatorHost, RuntimeConfigDefaults.LocatorHost),
+                LocatorParameters = new Alpha.LocatorParameters
+                {
+                    PlayerIdentity = new Alpha.PlayerIdentityCredentials
+                    {
+                        PlayerIdentityToken = CommandLineUtility.GetCommandLineValue(
+                            commandLineArgs, RuntimeConfigNames.PlayerIdentityToken, string.Empty),
+                        LoginToken = CommandLineUtility.GetCommandLineValue(
+                            commandLineArgs, RuntimeConfigNames.LoginToken, string.Empty)
+                    },
+                },
+            };
         }
     }
 }