+++ workers/unity/Packages/com.improbable.gdk.mobile/Worker/MobileWorkerConnector.cs
@@ -7,26 +7,44 @@ namespace Improbable.Gdk.Mobile
     {
         protected abstract string GetHostIp();
 
+        protected override ConnectionParameters GetConnectionParameters(string workerType, ConnectionService service)
+        {
+            return new ConnectionParameters
+            {
+                WorkerType = workerType,
+                Network =
+                {
+                    ConnectionType = NetworkConnectionType.Tcp,
+                    UseExternalIp = true,
+                },
+                EnableProtocolLoggingAtStartup = false,
+                DefaultComponentVtable = new ComponentVtable(),
+            };
+        }
+
         protected override ReceptionistConfig GetReceptionistConfig(string workerType)
         {
             return new ReceptionistConfig
             {
                 ReceptionistHost = GetHostIp(),
-                WorkerType = workerType,
-                WorkerId = CreateNewWorkerId(workerType),
-                UseExternalIp = true,
-                LinkProtocol = NetworkConnectionType.Tcp,
+                ReceptionistPort = RuntimeConfigDefaults.ReceptionistPort,
+                WorkerId = CreateNewWorkerId(workerType)
             };
         }
 
-        protected override LocatorConfig GetLocatorConfig(string workerType)
+        protected override LocatorConfig GetLocatorConfig()
         {
             throw new System.NotImplementedException("The locator flow is currently not available for mobile workers.");
         }
 
-        protected override bool ShouldUseLocator()
+        protected override AlphaLocatorConfig GetAlphaLocatorConfig()
         {
-            return false;
+            throw new System.NotImplementedException("The alpha locator flow is currently not available for mobile workers.");
+        }
+
+        protected override ConnectionService GetConnectionService()
+        {
+            return ConnectionService.Receptionist;
         }
     }
 }